Crate conduit

Structs

Response

TypeMap

Enums

Host

Method

Scheme

Traits

Handler
A Handler takes a request and returns a response or an error. By default, a bare function implements Handler.

Headers

Request

Type Definitions

Extensions
A Dictionary for extensions provided by the server or middleware