Trait conduit::Request

pub trait Request {
    fn http_version(&self) -> Version;
    fn conduit_version(&self) -> Version;
    fn method(&self) -> Method;
    fn scheme(&self) -> Scheme;
    fn host<'a>(&'a self) -> Host<'a>;
    fn virtual_root<'a>(&'a self) -> Option<&'a str>;
    fn path<'a>(&'a self) -> &'a str;
    fn query_string<'a>(&'a self) -> Option<&'a str>;
    fn remote_addr(&self) -> SocketAddr;
    fn content_length(&self) -> Option<u64>;
    fn headers<'a>(&'a self) -> &'a Headers;
    fn body<'a>(&'a mut self) -> &'a mut Read;
    fn extensions<'a>(&'a self) -> &'a Extensions;
    fn mut_extensions<'a>(&'a mut self) -> &'a mut Extensions;
}

Required Methods

fn http_version(&self) -> Version

The version of HTTP being used

fn conduit_version(&self) -> Version

The version of the conduit spec being used

fn method(&self) -> Method

The request method, such as GET, POST, PUT, DELETE or PATCH

fn scheme(&self) -> Scheme

The scheme part of the request URL

fn host<'a>(&'a self) -> Host<'a>

The host part of the requested URL

fn virtual_root<'a>(&'a self) -> Option<&'a str>

The initial part of the request URL's path that corresponds to a virtual root. This allows an application to have a virtual location that consumes part of the path.

fn path<'a>(&'a self) -> &'a str

The remainder of the path.

fn query_string<'a>(&'a self) -> Option<&'a str>

The portion of the request URL that follows the "?"

fn remote_addr(&self) -> SocketAddr

The remote IP address of the client or the last proxy that sent the request.

fn content_length(&self) -> Option<u64>

The byte-size of the body, if any

fn headers<'a>(&'a self) -> &'a Headers

The request's headers, as conduit::Headers.

fn body<'a>(&'a mut self) -> &'a mut Read

A Reader for the body of the request

fn extensions<'a>(&'a self) -> &'a Extensions

A readable map of extensions

fn mut_extensions<'a>(&'a mut self) -> &'a mut Extensions

A mutable map of extensions